Propagate topics, categories, and access levels correctly
